Fix graduation date
I noticed something I didn't like about how my graduation date was formatting
and realized I had done my dates in a way I don't like.  When I originally hand
typed JSON data that backs my resume, I used ISO-8601 format for the dates
because it was easy to type in by hand.  This didn't show any signs of a
problem in most of my dates because I gave a specific enough date for angular's
"date: longDate" to work out.  On my graduation date, however, I only supplied
the month and year (which is valid) and this caused angular to just display
"2007-12".

ISO-8601 isn't ideal for communicating a date between systems in my opinion,
because it is more complicated to parse/interpret than the milliseconds since
the epoch and it opens up ambiguity on things like specifying only the year
("2007" is valid ISO-8601).  My preference is to force the code creating the
date to be a little more explicit.

In this commit, I've changed the problem graduation_date to be a unix timestamp
(milliseconds since 1970), but left the other dates alone.  When transitioning
this code from a standalone html file to a proper webapp (html, js, css,
server, DB), I will address changing the data of the other dates.